+++ ProphetsWay.Hasher.Tests/HasherTests.cs
@@ -1,6 +1,5 @@
 using System.IO;
 using Xunit;
-using ProphetsWay.Utilities;
 using FluentAssertions;
 using System.Collections.Generic;
 
@@ -147,7 +146,7 @@ namespace ProphetsWay.Hasher.Tests
 #endif
 		public void TestGenerateHashFromFileName(string filename, string expectedHash, HashTypes hashType)
 		{
-			var hashResult = Utilities.Hasher.GenerateHash(filename, hashType);
+			var hashResult = Hasher.GenerateHash(filename, hashType);
 
 			hashResult.Should().Be(expectedHash);
 		}
@@ -157,38 +156,6 @@ namespace ProphetsWay.Hasher.Tests
 			var fi = new FileInfo(filename);
 			var fs = fi.Open(FileMode.Open, FileAccess.Read, FileShare.Read);
 			return fs;
-		}
-
-		[Theory]
-		[InlineData(TestFiles.TestFileA.Name, TestFiles.TestFileA.MD5, TestFiles.TestFileA.SHA1, TestFiles.TestFileA.SHA256, TestFiles.TestFileA.SHA512)]
-		[InlineData(TestFiles.TestFileB.Name, TestFiles.TestFileB.MD5, TestFiles.TestFileB.SHA1, TestFiles.TestFileB.SHA256, TestFiles.TestFileB.SHA512)]
-		[InlineData(TestFiles.TestFileC.Name, TestFiles.TestFileC.MD5, TestFiles.TestFileC.SHA1, TestFiles.TestFileC.SHA256, TestFiles.TestFileC.SHA512)]
-		public async void TestGenerateHashesAsyncFromStream(string filename, string expectedMD5, string expectedSHA1, string expectedSHA256, string expectedSHA512)
-		{
-			var fs = GetFileStream(filename);
-
-			var hashResult = await fs.GenerateHashesAsync();
-
-			hashResult.MD5.Should().Be(expectedMD5);
-			hashResult.SHA1.Should().Be(expectedSHA1);
-			hashResult.SHA256.Should().Be(expectedSHA256);
-			hashResult.SHA512.Should().Be(expectedSHA512);
-		}
-
-		[Theory]
-		[InlineData(TestFiles.TestFileA.Name, TestFiles.TestFileA.MD5, TestFiles.TestFileA.SHA1, TestFiles.TestFileA.SHA256, TestFiles.TestFileA.SHA512)]
-		[InlineData(TestFiles.TestFileB.Name, TestFiles.TestFileB.MD5, TestFiles.TestFileB.SHA1, TestFiles.TestFileB.SHA256, TestFiles.TestFileB.SHA512)]
-		[InlineData(TestFiles.TestFileC.Name, TestFiles.TestFileC.MD5, TestFiles.TestFileC.SHA1, TestFiles.TestFileC.SHA256, TestFiles.TestFileC.SHA512)]
-		public void TestGenerateHashesFromStream(string filename, string expectedMD5, string expectedSHA1, string expectedSHA256, string expectedSHA512)
-		{
-			var fs = GetFileStream(filename);
-
-			var hashResult = fs.GenerateHashes();
-
-			hashResult.MD5.Should().Be(expectedMD5);
-			hashResult.SHA1.Should().Be(expectedSHA1);
-			hashResult.SHA256.Should().Be(expectedSHA256);
-			hashResult.SHA512.Should().Be(expectedSHA512);
 		}
 
 		private HashTypes GetRolledUpHashTypes(HashTypes[] hashTypes)