+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v1.1.0
 ### Added Algorithms
 Added support for Hashing Algorithms SHA384 and RIPEMD160.  You can now specify them using the HashTypes enum.
+Created a new way to ask for multiple hash checksums against a single stream.  Depricated old methods that returned
+HashCollection results.  Those methods will be removed in a new version.
 
 
 # v1.0.1